api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supertokens-python==0.22.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
